Refactor low-level API into reusable object model
