+++ aeon/io/api.py
@@ -1,14 +1,12 @@
 import os
 import glob
-import math
 import bisect
 import datetime
 import pandas as pd
-import numpy as np
+from pathlib import Path
 
 """The duration of each acquisition chunk, in whole hours."""
 CHUNK_DURATION = 1
-_SECONDS_PER_TICK = 32e-6
 
 def aeon(seconds):
     """Converts a Harp timestamp, in seconds, to a datetime object."""
@@ -72,22 +70,7 @@ def chunk_filter(files, timefilter):
         matches.append(file)
     return matches
 
-def chunk_glob(path, device, prefix=None, extension='*.csv'):
-    '''
-    Finds matching files in the specified root path.
-
-    :param str path: The root path where all the epoch data is stored.
-    :param str, device: The device name prefix used to search for data files.
-    :param str, optional prefix: The pathname prefix used to search for data files.
-    :param str, optional extension: The optional extension pattern used to search for data files.
-    :return: A list of file names for chunks in the specified path.
-    '''
-    pathname = path + "/**/" + device + "/" + prefix + extension
-    files = glob.glob(pathname)
-    return files
-
-def load(path, reader, device, prefix=None, extension="*.csv",
-         start=None, end=None, time=None, tolerance=None):
+def load(root, reader, start=None, end=None, time=None, tolerance=None):
     '''
     Extracts data from matching files in the specified root path, sorted chronologically,
     containing device and/or epoch metadata for the Experiment 0.2 arena. If no prefix is
@@ -105,21 +88,12 @@ def load(path, reader, device, prefix=None, extension="*.csv",
     The maximum distance between original and new timestamps for inexact matches.
     :return: A pandas data frame containing epoch event metadata, sorted by priority and time.
     '''
-    if prefix is None:
-        prefix = ""
-
-    if isinstance(path, str):
-        files = chunk_glob(path, device, prefix, extension)
-    else:
-        files = []
-        fileset = set()
-        datasets = map(lambda dpath:chunk_glob(dpath, device, prefix, extension), path)
-        for dataset in datasets:
-            for fname in dataset:
-                fkey = os.path.split(fname)[1]
-                if not fkey in fileset:
-                    files.append(fname)
-    files.sort()
+    files = set()
+    if isinstance(root, str):
+        root = [root]
+    for path in root:
+        files.update(Path(path).glob(f"**/**/{reader.name}*.{reader.extension}"))
+    files = sorted(files)
 
     if time is not None:
         # ensure input is converted to timestamp series
@@ -134,9 +108,9 @@ def load(path, reader, device, prefix=None, extension="*.csv",
         for key,values in time.groupby(by=chunk):
             i = bisect.bisect_left(filetimes, key)
             if i < len(filetimes):
-                frame = reader(files[i])
+                frame = reader.read(files[i])
             else:
-                frame = reader(None)
+                frame = reader.read(None)
             data = frame.reset_index()
             data.set_index('time', drop=False, inplace=True)
             data = data.reindex(values, method='pad', tolerance=tolerance)
@@ -144,7 +118,7 @@ def load(path, reader, device, prefix=None, extension="*.csv",
             if missing > 0 and i > 0:
                 # expand reindex to allow adjacent chunks
                 # to fill missing values
-                previous = reader(files[i-1])
+                previous = reader.read(files[i-1])
                 data = pd.concat([previous, frame])
                 data = data.reindex(values, method='pad', tolerance=tolerance)
             else:
@@ -152,7 +126,7 @@ def load(path, reader, device, prefix=None, extension="*.csv",
             dataframes.append(data)
 
         if len(dataframes) == 0:
-            return reader(None)
+            return reader.read(None)
             
         return pd.concat(dataframes)
 
@@ -163,437 +137,19 @@ def load(path, reader, device, prefix=None, extension="*.csv",
         timefilter = None
 
     if len(files) == 0:
-        return reader(None)
+        return reader.read(None)
 
-    data = pd.concat([reader(file) for file in files])
+    data = pd.concat([reader.read(file) for file in files])
     if timefilter is not None:
         try:
             return data.loc[start:end]
         except KeyError:
             import warnings
             if not data.index.has_duplicates:
-                warnings.warn('data index for {0} contains out-of-order timestamps!'.format(device))
+                warnings.warn('data index for {0} contains out-of-order timestamps!'.format(reader.name))
                 data = data.sort_index()
             else:
-                warnings.warn('data index for {0} contains duplicate keys!'.format(device))
+                warnings.warn('data index for {0} contains duplicate keys!'.format(reader.name))
                 data = data[~data.index.duplicated(keep='first')]
             return data.loc[start:end]
-    return data
-
-def chunkreader(file):
-    """Reads acquisition chunk information for the specified file."""
-    names = ['time', 'path']
-    if file is None:
-        return pd.DataFrame(columns=names[1:], index=pd.DatetimeIndex([]))
-    chunk = chunk_key(file)
-    data = pd.DataFrame({ names[0]: [chunk], names[1]: [file] })
-    data.set_index('time', inplace=True)
-    return data
-
-def chunkdata(path, device, extension='*.csv', start=None, end=None, time=None, tolerance=None):
-    '''
-    Extracts all acquisition chunk information from the specified root path, sorted chronologically,
-    indicating recorded acquisition chunks for the specified device in the arena.
-
-    :param str path: The root path where all the data is stored.
-    :param str, device: The device name prefix used to search for acquisition chunk files.
-    :param str, optional extension: The optional extension pattern used to search for chunk files.
-    :param datetime, optional start: The left bound of the time range to extract.
-    :param datetime, optional end: The right bound of the time range to extract.
-    :param datetime, optional time: An object or series specifying the timestamps to extract.
-    :param datetime, optional tolerance:
-    The maximum distance between original and new timestamps for inexact matches.
-    :return: A pandas data frame containing acquisition chunk information, sorted by time.
-    '''
-    return load(
-        path,
-        chunkreader,
-        device,
-        extension=extension,
-        start=start,
-        end=end,
-        time=time,
-        tolerance=tolerance)
-
-def subjectreader(file):
-    """Reads subject metadata from the specified file."""
-    names = ['time','id','weight','event']
-    if file is None:
-        return pd.DataFrame(columns=names[1:], index=pd.DatetimeIndex([]))
-    data = pd.read_csv(file, header=0, names=names)
-    data['time'] = aeon(data['time'])
-    data.set_index('time', inplace=True)
-    return data
-
-def subjectdata(path, start=None, end=None, time=None, tolerance=None):
-    '''
-    Extracts all subject state metadata from the specified root path, sorted
-    chronologically, indicating enter and exit times of subjects in the
-    Experiment 0.2 arena.
-
-    :param str path: The root path where all the subject state metadata is stored.
-    :param datetime, optional start: The left bound of the time range to extract.
-    :param datetime, optional end: The right bound of the time range to extract.
-    :param datetime, optional time: An object or series specifying the timestamps to extract.
-    :param datetime, optional tolerance:
-    The maximum distance between original and new timestamps for inexact matches.
-    :return: A pandas data frame containing subject state metadata, sorted by time.
-    '''
-    return load(
-        path,
-        subjectreader,
-        device='ExperimentalMetadata',
-        prefix='ExperimentalMetadata_SubjectState',
-        extension="*.csv",
-        start=start,
-        end=end,
-        time=time,
-        tolerance=tolerance)
-
-def logreader(file):
-    """Reads message log data from the specified file."""
-    names = ['time','priority','type','message']
-    if file is None:
-        return pd.DataFrame(columns=names[1:], index=pd.DatetimeIndex([]))
-    data = pd.read_csv(file, header=0, names=names)
-    data['time'] = aeon(data['time'])
-    data.set_index('time', inplace=True)
-    return data
-
-def logdata(path, start=None, end=None, time=None, tolerance=None):
-    '''
-    Extracts log message data from the specified root path, sorted chronologically,
-    containing alerts and manual annotations for the Experiment 0.2 arena.
-
-    :param str path: The root path where all the epoch data is stored.
-    :param datetime, optional start: The left bound of the time range to extract.
-    :param datetime, optional end: The right bound of the time range to extract.
-    :param datetime, optional time: An object or series specifying the timestamps to extract.
-    :param datetime, optional tolerance:
-    The maximum distance between original and new timestamps for inexact matches.
-    :return: A pandas data frame containing log message data, sorted by time.
-    '''
-    return load(
-        path,
-        logreader,
-        device='ExperimentalMetadata',
-        prefix='ExperimentalMetadata_MessageLog',
-        extension="*.csv",
-        start=start,
-        end=end,
-        time=time,
-        tolerance=tolerance)
-
-def videoreader(file):
-    """Reads video metadata from the specified file."""
-    names = ['time','hw_counter','hw_timestamp']
-    if file is None:
-        return pd.DataFrame(columns=['frame']+names[1:], index=pd.DatetimeIndex([]))
-    data = pd.read_csv(file, header=0, names=names)
-    data.insert(loc=1, column='frame', value=data.index)
-    data['time'] = aeon(data['time'])
-    data['path'] = os.path.splitext(file)[0] + '.avi'
-    data['epoch'] = file.rsplit(os.sep, maxsplit=3)[1]
-    data.set_index('time', inplace=True)
-    return data
-
-def videodata(path, device, start=None, end=None, time=None, tolerance=None):
-    '''
-    Extracts all video metadata from the specified root path, sorted chronologically,
-    indicating synchronized trigger frame times for cameras in the Experiment 0 arena.
-
-    :param str path: The root path where all the video data is stored.
-    :param str, device: The device prefix used to search for video files.
-    :param datetime, optional start: The left bound of the time range to extract.
-    :param datetime, optional end: The right bound of the time range to extract.
-    :param datetime, optional time: An object or series specifying the timestamps to extract.
-    :param datetime, optional tolerance:
-    The maximum distance between original and new timestamps for inexact matches.
-    :return: A pandas data frame containing frame event metadata, sorted by time.
-    '''
-    return load(
-        path,
-        videoreader,
-        device=device,
-        extension="*.csv",
-        start=start,
-        end=end,
-        time=time,
-        tolerance=tolerance)
-
-def videoclip(path, device, start=None, end=None):
-    '''
-    Extracts information about a continuous segment of video, possibly stored across
-    multiple video files. For each video file covering the segment, a row is returned
-    containing the path, start frame, and duration of the segment stored in that file.
-
-    :param str path: The root path where all the video data is stored.
-    :param str, device: The device prefix used to search for video files.
-    :param datetime, optional start: The left bound of the time range to extract.
-    :param datetime, optional end: The right bound of the time range to extract.
-    :return: A pandas data frame containing video clip storage information.
-    '''
-    framedata = videodata(path, device, start=start, end=end)
-    if len(framedata) == 0:
-        return pd.DataFrame(columns=['start','duration'], index=pd.DatetimeIndex([]))
-    videoclips = framedata.groupby('path')
-    startframe = videoclips.frame.min().rename('start')
-    duration = (videoclips.frame.max() - startframe).rename('duration')
-    return pd.concat([startframe, duration], axis=1)
-
-"""Maps Harp payload types to numpy data type objects."""
-payloadtypes = {
-    1 : np.dtype(np.uint8),
-    2 : np.dtype(np.uint16),
-    4 : np.dtype(np.uint32),
-    8 : np.dtype(np.uint64),
-    129 : np.dtype(np.int8),
-    130 : np.dtype(np.int16),
-    132 : np.dtype(np.int32),
-    136 : np.dtype(np.int64),
-    68 : np.dtype(np.float32)
-}
-
-"""Map of Harp device registers used in the Experiment 1 arena."""
-harpregisters = {
-    ('Patch', 90) : ['angle, intensity'],   # wheel encoder
-    ('Patch', 35) : ['bitmask'],            # trigger pellet delivery
-    ('Patch', 32) : ['bitmask'],            # pellet detected by beam break
-    ('VideoController', 68) : ['pwm_mask'], # camera trigger times (top and side)
-    ('PositionTracking', 200) : ['x', 'y', 'angle', 'major', 'minor', 'area'],
-    ('WeightScale', 200) : ['value', 'stable']
-}
-
-def harpreader(file, names=None):
-    '''
-    Reads Harp data from the specified file.
-    
-    :param str file: The path to a Harp binary file.
-    :param str or array-like names: The optional column labels to use for the data.
-    :return: A pandas data frame containing harp event data, sorted by time.
-    '''
-    if file is None:
-        return pd.DataFrame(columns=names, index=pd.DatetimeIndex([]))
-
-    data = np.fromfile(file, dtype=np.uint8)
-    if len(data) == 0:
-        return pd.DataFrame(columns=names, index=pd.DatetimeIndex([]))
-
-    stride = data[1] + 2
-    length = len(data) // stride
-    payloadsize = stride - 12
-    payloadtype = payloadtypes[data[4] & ~0x10]
-    elementsize = payloadtype.itemsize
-    payloadshape = (length, payloadsize // elementsize)
-    seconds = np.ndarray(length, dtype=np.uint32, buffer=data, offset=5, strides=stride)
-    ticks = np.ndarray(length, dtype=np.uint16, buffer=data, offset=9, strides=stride)
-    seconds = ticks * _SECONDS_PER_TICK + seconds
-    payload = np.ndarray(
-        payloadshape,
-        dtype=payloadtype,
-        buffer=data, offset=11,
-        strides=(stride, elementsize))
-    time = aeon(seconds)
-    time.name = 'time'
-
-    if names is not None and payloadshape[1] < len(names):
-        data = pd.DataFrame(payload, index=time, columns=names[:payloadshape[1]])
-        data[names[payloadshape[1]:]] = math.nan
-        return data
-    else:
-        return pd.DataFrame(payload, index=time, columns=names)
-
-def harpdata(path, device, register, names=None, start=None, end=None, time=None, tolerance=None):
-    '''
-    Extracts all harp data from the specified root path, sorted chronologically,
-    for an individual register acquired from a device in the Experiment 0 arena.
-
-    :param str path: The root path where all the data is stored.
-    :param str device: The device name used to search for data files.
-    :param int register: The register number to extract data for.
-    :param str or array-like names: The optional column labels to use for the data.
-    :param datetime, optional start: The left bound of the time range to extract.
-    :param datetime, optional end: The right bound of the time range to extract.
-    :param datetime, optional time: An object or series specifying the timestamps to extract.
-    :param datetime, optional tolerance:
-    The maximum distance between original and new timestamps for inexact matches.
-    :return: A pandas data frame containing harp event data, sorted by time.
-    '''
-    return load(
-        path,
-        lambda file: harpreader(file, names),
-        device=device,
-        prefix="{0}_{1}".format(device, register),
-        extension="*.bin",
-        start=start,
-        end=end,
-        time=time,
-        tolerance=tolerance)
-
-def encoderdata(path, device, start=None, end=None, time=None, tolerance=None):
-    '''
-    Extracts all encoder data from the specified root path, sorted chronologically,
-    from the specified patch controller in the Experiment 0 arena.
-
-    :param str path: The root path where all the data is stored.
-    :param str device: The device name used to search for data files.
-    :param datetime, optional start: The left bound of the time range to extract.
-    :param datetime, optional end: The right bound of the time range to extract.
-    :return: A pandas data frame containing encoder data, sorted by time.
-    '''
-    return harpdata(
-        path,
-        device, register=90,
-        names=['angle', 'intensity'],
-        start=start, end=end,
-        time=time, tolerance=tolerance)
-
-def pelletdata(path, device, start=None, end=None, time=None, tolerance=None):
-    '''
-    Extracts all pellet event data from the specified root path, sorted chronologically,
-    indicating when delivery of a pellet was triggered and when pellets were detected
-    by the feeder beam break.
-
-    :param str path: The root path where all the epoch data is stored.
-    :param str device: The device name used to search for data files.
-    :param datetime, optional start: The left bound of the time range to extract.
-    :param datetime, optional end: The right bound of the time range to extract.
-    :return: A pandas data frame containing pellet event data, sorted by time.
-    '''
-    command = harpdata(path, device, register=35, names=['bitmask'],
-                       start=start, end=end, time=time, tolerance=tolerance)
-    beambreak = harpdata(path, device, register=32, names=['bitmask'],
-                         start=start, end=end, time=time, tolerance=tolerance)
-    command = command[command.bitmask == 0x80]
-    beambreak = beambreak[beambreak.bitmask == 0x20]
-    command['event'] = 'TriggerPellet'
-    beambreak['event'] = 'PelletDetected'
-    events = pd.concat([command.event, beambreak.event]).sort_index()
-    return pd.DataFrame(events)
-
-def weightdata(path, device='Nest', start=None, end=None, time=None, tolerance=None):
-    '''
-    Extracts weight data from the specified root path, sorted chronologically,
-    for the specified electronic weighing device in the Experiment arena.
-
-    :param str path: The root path where all the epoch data is stored.
-    :param str patch: The device name used to search for data files.
-    :param datetime, optional start: The left bound of the time range to extract.
-    :param datetime, optional end: The right bound of the time range to extract.
-    :param datetime, optional time: An object or series specifying the timestamps to extract.
-    :param datetime, optional tolerance:
-    The maximum distance between original and new timestamps for inexact matches.
-    :return: A pandas data frame, sorted by time, containing absolute weight in grams,
-    and a value indicating whether the reading is stable.
-    '''
-    return harpdata(
-        path,
-        device,
-        register=200,
-        names=['value','stable'],
-        start=start,
-        end=end,
-        time=time,
-        tolerance=tolerance)
-
-def patchreader(file):
-    """Reads patch state metadata from the specified file."""
-    names = ['time','threshold','d1','delta']
-    if file is None:
-        return pd.DataFrame(columns=names[1:], index=pd.DatetimeIndex([]))
-    data = pd.read_csv(file, header=0, names=names)
-    data['time'] = aeon(data['time'])
-    data.set_index('time', inplace=True)
-    return data
-
-def patchdata(path, patch, start=None, end=None, time=None, tolerance=None):
-    '''
-    Extracts patch metadata from the specified root path, sorted chronologically,
-    indicating wheel threshold, d1 and delta state changes in the arena patch.
-
-    :param str path: The root path where all the epoch data is stored.
-    :param str patch: The patch name used to search for data files.
-    :param datetime, optional start: The left bound of the time range to extract.
-    :param datetime, optional end: The right bound of the time range to extract.
-    :param datetime, optional time: An object or series specifying the timestamps to extract.
-    :param datetime, optional tolerance:
-    The maximum distance between original and new timestamps for inexact matches.
-    :return: A pandas data frame containing patch state metadata, sorted by time.
-    '''
-    return load(
-        path,
-        patchreader,
-        device=patch,
-        prefix='{0}_State'.format(patch),
-        extension="*.csv",
-        start=start,
-        end=end,
-        time=time,
-        tolerance=tolerance)
-
-def positiondata(path, device='CameraTop', start=None, end=None, time=None, tolerance=None):
-    '''
-    Extracts all position data from the specified root path, sorted chronologically,
-    for the specified camera in the Experiment 0 arena.
-
-    :param str path: The root path where all the data is stored.
-    :param str device: The device name used to search for data files.
-    :param datetime, optional start: The left bound of the time range to extract.
-    :param datetime, optional end: The right bound of the time range to extract.
-    :param datetime, optional time: An object or series specifying the timestamps to extract.
-    :param datetime, optional tolerance:
-    The maximum distance between original and new timestamps for inexact matches.
-    :return: A pandas data frame containing position data, sorted by time.
-    '''
-    return harpdata(
-        path,
-        device, register=200,
-        names=['x', 'y', 'angle', 'major', 'minor', 'area', 'id'],
-        start=start,
-        end=end,
-        time=time,
-        tolerance=tolerance)
-
-def distancetravelled(angle, radius=4.0):
-    '''
-    Calculates the total distance travelled on the wheel, by taking into account
-    its radius and the total number of turns in both directions across time.
-
-    :param Series angle: A series of magnetic encoder measurements.
-    :param float radius: The radius of the wheel, in metric units.
-    :return: The total distance travelled on the wheel, in metric units.
-    '''
-    maxvalue = int(np.iinfo(np.uint16).max >> 2)
-    jumpthreshold = maxvalue // 2
-    turns = angle.astype(int).diff()
-    clickup = (turns < -jumpthreshold).astype(int)
-    clickdown = (turns > jumpthreshold).astype(int) * -1
-    turns = (clickup + clickdown).cumsum()
-    distance = 2 * np.pi * radius * (turns + angle / maxvalue)
-    distance = distance - distance[0]
-    return distance
-
-def subjectvisit(data):
-    '''
-    Computes duration, enter and exit times for a subject visit. Allows for missing data by
-    trying to match subject enter times with subsequent exit times. If the match fails,
-    subject exit metadata is filled with NaN. Any additional metadata columns in the data
-    frame will be paired and included in the output.
-
-    :param DataFrame data: A pandas data frame containing subject enter and exit events.
-    :return: A pandas data frame containing duration and metadata for each visit.
-    '''
-    enter = data.event == 'Enter'
-    exit = data[enter.shift(1, fill_value=False)].reset_index()
-    enter = data[enter].reset_index()
-    data = enter.join(exit, lsuffix='_enter', rsuffix='_exit')
-    valid_subjects = (data.id_enter == data.id_exit) & (data.event_exit == 'Exit')
-    if ~valid_subjects.any():
-        data_types = data.dtypes
-        data.loc[~valid_subjects, [name for name in data.columns if '_exit' in name]] = None
-        data = data.astype(data_types)
-    data['duration'] = data.time_exit - data.time_enter
-    data.rename({ 'time_enter':'enter', 'id_enter':'id', 'time_exit':'exit'}, axis=1, inplace=True)
-    data = data[['id'] + [name for name in data.columns if '_' in name] + ['enter', 'exit', 'duration']]
-    data.drop(['id_exit', 'event_enter', 'event_exit'], axis=1, inplace=True)
     return data
